IntigrityLens ai/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.29
aiosqlite==0.20.0
python-multipart==0.0.9
reportlab==4.1.0
python-dotenv==1.0.1
